tabbar: fix error of circular reference
If Tabbar PropTypes is at shapes, it needs a reference to TabItem. But
it's referred by TabItem, because it contains its types and defaults
too.

Put types and default back to its files, and left just the common
parts at shapes.

TabItem children spec removed because of error of unused + and it can
be anything